Modularized ovsp4rt.cc
- Separated target-independent functions from functions that are
  specific to DPDK or ES2K.

- Separated public functions (ones exposed to openvswitch through
  ovs-p4rt.h) from private functions (ones internal to ovs-p4rt).
