+++ ovs-p4rt/sidecar/CMakeLists.txt
@@ -6,10 +6,12 @@
 
 add_library(ovs_sidecar_o OBJECT
     ovsp4rt.cc
+    ovsp4rt.h
     ovsp4rt_session.cc
     ovsp4rt_session.h
     ovsp4rt_credentials.cc
     ovsp4rt_credentials.h
+    ovsp4rt_public.cc
     ovsp4rt_utils.cc
     ovsp4rt_utils.h
 )
@@ -21,6 +23,7 @@ elseif(ES2K_TARGET)
 endif()
 
 target_include_directories(ovs_sidecar_o PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}
     ${OVS_INCLUDE_DIR}
     ${STRATUM_SOURCE_DIR}
     ${PB_OUT_DIR}
